Use ping option as the timeout value for websocket connections
The `ping` is meant to specify how often to send a keep alive ping
to active websockets, to avoid idle connections being automatically
killed by services like Heroku. This may or may not actually be
working, I haven't figured that part out. However, there seems to be
some other kind of ping going on, where every N seconds (N being the
configured timeout interval for your Faye server), the server sends
a `/meta/connect` frame to the client, which acknowledges it with a
frame of its own.

If you initialize your faye server with different `ping` and `timeout`
values, the `timeout` value is used for that `/meta/connect` polling,
which is undesirable when attempting to minimize the number of pings
and maximize battery life.

This change makes it so that the `/meta/connect` poll happens every
`ping` seconds for websocket connections, and every `timeout` seconds
for any other transport type. If the `ping` is unset, the `timeout`
is used in all cases.